Update PhysCrate.cc
We would like to throw errors if we cannot set up an A2795 board we expect to exist, however this is complicated by the way PhysCrate runs over the link. PhysCrate right now checks node on the link in a brute force manner with no concern for how many boards the crate should have. This means we cannot throw an error on a failed connection because we have some links which have fewer than the hard coded max of 5.
A smarter way to handle this is to pass the number of boards expected on each link and then throw an error if we try to find a board and can't. This also requires adding a vector to the crate configuration, but since we should know the number of boards in each crate when starting the run it shouldn't be too onerous.